Move evolution code into separate module
